pyproject.toml
[tool.poetry]
name = "project-buyme"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "project_buyme"}]

[tool.poetry.dependencies]
python = "^3.7"
allure-pytest = "^2.13.2"
selenium = "^4.11.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
